chore: push correct tags to ECR repository
The workflow only pushes the `latest` tag before deploying to pre-prod,
but the image is expected to be tagged with the commit sha. Add the
`--all-tags` flag to the Docker push command.

The `stable` tag is used in the microservice template project but is not
used in any `tis-trainee` workflows, remove it for consistency and use
the commit sha directly instead.

TIS21-3337
TIS21-3343