feat: update account email when TIS changes
Update the email address associated with a user account when the TIS
contact details are changed.
The email should only be updated if it is not in use by any other user.

Exceptions should be throw in two cases
 - Where multiple accounts exist for a given PersonId
 - Where the new email address is already in use

Cache the users retrieved from cognito as the lookup is costly.

TIS21-5907
TIS21-5940