fix: tracing configuration
The AWS X-Ray interceptor is configured to wrap Service classes,
but when the service class is called from an SQS Listener the trace
segments are not instrumented which causes errors in the logs.

Remove the wrapping of the service beans, leaving just the resource
beans. The service calls will still appear in the traces but they will
not attempt to be wrapped.

NO-TICKET