fix: include user IDs in multiple account error
When attempting to update an email an exception is throw if multiple
accounts are found for a given trainee ID.
The current expection message says how many were found, but it is
challenging to query for `custom:tisId` to identify problem accounts.

Update ContactDetailsListener to include the found user IDs in the error
message. This should allow us to take appropriate action when these
sorts of errors occur.

NO-TICKET