feat: delete duplicate accounts on email update
When contact details are updated the trainee's account is updated to
change the log-in email. Currently this update is rejected if there are
multiple accounts for the trainee, based on the Person ID.
Update the ContactDetailsListener and UserAccountService to attempt to
delete the duplicate accounts, before finalising the email update.

The initial rule for deletion is to retain any account that matches the
current contact email address, deleting the other duplicates.
Further rules will be added over time once the remaining scenarios can
be identified.

TIS21-7288