feat: update custom MFA attribute on lookup
Refactor the UserAccountService to split Cognito API calls in to a new
CognitoService. This should be used to wrap calls to Cognito to allow
additional functionality to be rolled in to the calls.

Create a `getUserDetails` method which wraps `AdminGetUser` and attempts
to get `ListUsers` as an alternative which does not increase MAU costs.
As the MFA settings are not available via `ListUsers` they should be
retrieved from `AdminGetUser` if not known and then populated in to a
`custom:mfaType` attribute to make it available to `ListUsers` next
time.

An attempt was made to add caching, but further refactoring is needed to
allow consistent cache eviction. The current mix of `email` and `sub`
lookups would most likely lead to stale data in the cache without very
careful manual eviction.

As the MFA details are more readily available the reset metrics
publishing should be fixed to send the accurate MFA value instead of
`NO_MFA`.

TIS21-7256
TIS21-7276
TIS21-7393