fix: catch UserNotFoundException finding dupes
When trying to de-duplicate accounts, if the new TIS email does not
match an existing account an exception is thrown. This scenario should
be handled more gracefully, with the exception caught and treated as no
main account found.

The auth events should still be checked so that successless accounts can
be identified.

TIS21-7288